chore: addressing PR comments
