+++ src/@types/index.ts
@@ -2,7 +2,6 @@ import './global';
 
 export * from '@reduxjs/toolkit';
 export * from './api';
-export * from './events';
 export * from './i18n';
 export * from './model';
 export * from './new';
