+++ src/store/eventing/index.ts
@@ -1,1 +1,2 @@
 export { default } from './middleware';
+export * from './types';
