+++ src/store/eventing/middleware.ts
@@ -2,7 +2,7 @@ import noop from 'lodash/noop';
 import { Action, Dispatch, Middleware, MiddlewareAPI } from '@reduxjs/toolkit';
 import { saveAnnotationAction } from '../annotations/actions';
 import { handleCreateErrorEvents, handleCreatePendingEvents, handleCreateSuccessEvents } from './create';
-import { EventHandlerMap } from '../../@types';
+import { EventHandlerMap } from './types';
 
 // Array of event handlers based on redux action. To add handling for new events add an entry keyed by action
 const eventHandlers: EventHandlerMap = {
@@ -14,6 +14,8 @@ const eventHandlers: EventHandlerMap = {
 function getEventingMiddleware(handlers: EventHandlerMap = eventHandlers): Middleware {
     return (store: MiddlewareAPI) => (next: Dispatch) => (action: Action): Action => {
         const { type } = action;
+        // Retrieve the prevState as well as the nextState after the action has modified the state in order
+        // to pass all that information to the event handler
         const prevState = store.getState();
         const result = next(action);
         const nextState = store.getState();
