+++ src/store/eventing/types.ts
@@ -1,7 +1,9 @@
-import { Action } from '@reduxjs/toolkit';
+import { Action, SerializedError } from '@reduxjs/toolkit';
 import { ApplicationState } from '../types';
+import { Annotation, NewAnnotation } from '../../@types';
 
 export enum Event {
+    ANNOTATOR = 'annotatorevent',
     CREATE_ANNOTATION = 'annotationCreate',
 }
 
@@ -15,10 +17,26 @@ export interface Metadata {
     status: Status;
 }
 export interface ActionEvent {
-    annotation: object | undefined;
+    annotation: Annotation | undefined;
     error: Error | undefined;
     meta: Metadata;
 }
 
-export type EventHandler = (prevState: ApplicationState, nextState: ApplicationState, action: Action) => void;
+export interface ThunkMeta<M> {
+    arg: M;
+    requestId: string;
+}
+
+export interface AsyncAction extends Action {
+    error?: SerializedError;
+    meta?: ThunkMeta<NewAnnotation>;
+    payload?: Annotation;
+}
+
+export type EventHandler = (
+    prevState: ApplicationState,
+    nextState: ApplicationState,
+    action: Action | AsyncAction,
+) => void;
+
 export type EventHandlerMap = Record<Action['type'], EventHandler>;
